Refactored and removed unnecessary interface
