--- src/main/java/jenkins/plugins/gerrit/client/GerritClient.java
@@ -1,10 +1,0 @@
-package jenkins.plugins.gerrit.client;
-
-import com.google.gerrit.extensions.api.changes.Changes;
-import jenkins.plugins.gerrit.GerritURI;
-
-public interface GerritClient {
-    Changes changes();
-
-    String getScheme();
-}
